+++ src/Storage/Storage.Management/StorageAccount/NewAzureStorageAccount.cs
@@ -428,6 +428,23 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         [Parameter(
             Mandatory = false,
+            HelpMessage = "Specifies if managed identities can access SMB shares using OAuth. The default interpretation is false for this property.")]
+        [ValidateNotNullOrEmpty]
+        public bool EnableSmbOAuth
+        {
+            get
+            {
+                return enableSmbOAuth.Value;
+            }
+            set
+            {
+                enableSmbOAuth = value;
+            }
+        }
+        private bool? enableSmbOAuth = null;
+
+        [Parameter(
+            Mandatory = false,
             HelpMessage = "Specifies the Active Directory account type for Azure Storage. Possible values include: 'User', 'Computer'.",
             ParameterSetName = ActiveDirectoryDomainServicesForFileParameterSet)]
         [PSArgumentCompleter("User", "Computer")]
@@ -769,6 +786,19 @@ namespace Microsoft.Azure.Commands.Management.Storage
                 }
                 createParameters.AzureFilesIdentityBasedAuthentication.DefaultSharePermission = this.DefaultSharePermission;
             }
+
+            if (this.enableSmbOAuth != null)
+            {
+                if (createParameters.AzureFilesIdentityBasedAuthentication == null)
+                {
+                    createParameters.AzureFilesIdentityBasedAuthentication = new AzureFilesIdentityBasedAuthentication
+                    {
+                        DirectoryServiceOptions = DirectoryServiceOptions.None
+                    };
+                }
+                createParameters.AzureFilesIdentityBasedAuthentication.SmbOAuthSettings = new SmbOAuthSettings(this.enableSmbOAuth.Value);
+            }
+
             if (this.EnableLargeFileShare.IsPresent)
             {
                 createParameters.LargeFileSharesState = LargeFileSharesState.Enabled;